move v1 http config schemas to proper package
