+++ http/schema.go
@@ -1,37 +1,39 @@
-package schema
+package http
 
-func V1TLSSubformOptions() []Field {
-	return []Field{
+import "github.com/grafana/alerting/receivers/schema"
+
+func V1TLSSubformOptions() []schema.Field {
+	return []schema.Field{
 		{
 			Label:        "Disable certificate verification",
-			Element:      ElementTypeCheckbox,
+			Element:      schema.ElementTypeCheckbox,
 			Description:  "Do not verify the server's certificate chain and host name.",
 			PropertyName: "insecureSkipVerify",
 			Required:     false,
 		},
 		{
 			Label:        "CA Certificate",
-			Element:      ElementTypeTextArea,
+			Element:      schema.ElementTypeTextArea,
 			Description:  "Certificate in PEM format to use when verifying the server's certificate chain.",
-			InputType:    InputTypeText,
+			InputType:    schema.InputTypeText,
 			PropertyName: "caCertificate",
 			Required:     false,
 			Secure:       true,
 		},
 		{
 			Label:        "Client Certificate",
-			Element:      ElementTypeTextArea,
+			Element:      schema.ElementTypeTextArea,
 			Description:  "Client certificate in PEM format to use when connecting to the server.",
-			InputType:    InputTypeText,
+			InputType:    schema.InputTypeText,
 			PropertyName: "clientCertificate",
 			Required:     false,
 			Secure:       true,
 		},
 		{
 			Label:        "Client Key",
-			Element:      ElementTypeTextArea,
+			Element:      schema.ElementTypeTextArea,
 			Description:  "Client key in PEM format to use when connecting to the server.",
-			InputType:    InputTypeText,
+			InputType:    schema.InputTypeText,
 			PropertyName: "clientKey",
 			Required:     false,
 			Secure:       true,
@@ -39,19 +41,19 @@ func V1TLSSubformOptions() []Field {
 	}
 }
 
-func V1ProxyOption() Field {
-	return Field{ // New in 12.1.
+func V1ProxyOption() schema.Field {
+	return schema.Field{ // New in 12.1.
 		Label:        "Proxy Config",
 		PropertyName: "proxy_config",
 		Description:  "Optional proxy configuration.",
-		Element:      ElementTypeSubform,
-		SubformOptions: []Field{
+		Element:      schema.ElementTypeSubform,
+		SubformOptions: []schema.Field{
 			{
 				Label:        "Proxy URL",
 				PropertyName: "proxy_url",
 				Description:  "HTTP proxy server to use to connect to the targets.",
-				Element:      ElementTypeInput,
-				InputType:    InputTypeText,
+				Element:      schema.ElementTypeInput,
+				InputType:    schema.InputTypeText,
 				Placeholder:  "https://proxy.example.com",
 				Required:     false,
 				Secure:       false,
@@ -61,7 +63,7 @@ func V1ProxyOption() Field {
 				Label:        "Proxy from environment",
 				PropertyName: "proxy_from_environment",
 				Description:  "Use environment HTTP_PROXY, HTTPS_PROXY and NO_PROXY to determine proxies.",
-				Element:      ElementTypeCheckbox,
+				Element:      schema.ElementTypeCheckbox,
 				Required:     false,
 				Secure:       false,
 			},
@@ -69,8 +71,8 @@ func V1ProxyOption() Field {
 				Label:        "No Proxy",
 				PropertyName: "no_proxy",
 				Description:  "Comma-separated list of addresses that should not use a proxy.",
-				Element:      ElementTypeInput,
-				InputType:    InputTypeText,
+				Element:      schema.ElementTypeInput,
+				InputType:    schema.InputTypeText,
 				Placeholder:  "example.com,1.2.3.4",
 				Required:     false,
 				Secure:       false,
@@ -79,8 +81,8 @@ func V1ProxyOption() Field {
 				Label:        "Proxy Connect Header",
 				PropertyName: "proxy_connect_header",
 				Description:  "Optional headers to send to proxies during CONNECT requests.",
-				Element:      ElementTypeKeyValueMap,
-				InputType:    InputTypeText,
+				Element:      schema.ElementTypeKeyValueMap,
+				InputType:    schema.InputTypeText,
 				Required:     false,
 				Secure:       false,
 			},
@@ -88,25 +90,25 @@ func V1ProxyOption() Field {
 	}
 }
 
-func V1HttpClientOption() Field {
-	return Field{ // New in 12.1.
+func V1HttpClientOption() schema.Field {
+	return schema.Field{ // New in 12.1.
 		Label:        "HTTP Config",
 		PropertyName: "http_config",
 		Description:  "Common HTTP client options.",
-		Element:      ElementTypeSubform,
-		SubformOptions: []Field{
+		Element:      schema.ElementTypeSubform,
+		SubformOptions: []schema.Field{
 			{ // New in 12.1.
 				Label:        "OAuth2",
 				PropertyName: "oauth2",
 				Description:  "OAuth2 configuration options",
-				Element:      ElementTypeSubform,
-				SubformOptions: []Field{
+				Element:      schema.ElementTypeSubform,
+				SubformOptions: []schema.Field{
 					{
 						Label:        "Token URL",
 						PropertyName: "token_url",
-						Element:      ElementTypeInput,
+						Element:      schema.ElementTypeInput,
 						Description:  "URL for the access token endpoint.",
-						InputType:    InputTypeText,
+						InputType:    schema.InputTypeText,
 						Required:     true,
 						Secure:       false,
 						Protected:    true,
@@ -114,25 +116,25 @@ func V1HttpClientOption() Field {
 					{
 						Label:        "Client ID",
 						PropertyName: "client_id",
-						Element:      ElementTypeInput,
+						Element:      schema.ElementTypeInput,
 						Description:  "Client ID to use when authenticating.",
-						InputType:    InputTypeText,
+						InputType:    schema.InputTypeText,
 						Required:     true,
 						Secure:       false,
 					},
 					{
 						Label:        "Client Secret",
 						PropertyName: "client_secret",
-						Element:      ElementTypeInput,
+						Element:      schema.ElementTypeInput,
 						Description:  "Client secret to use when authenticating.",
-						InputType:    InputTypeText,
+						InputType:    schema.InputTypeText,
 						Required:     true,
 						Secure:       true,
 					},
 					{
 						Label:        "Scopes",
 						PropertyName: "scopes",
-						Element:      ElementStringArray,
+						Element:      schema.ElementStringArray,
 						Description:  "Optional scopes to request when obtaining an access token.",
 						Required:     false,
 						Secure:       false,
@@ -140,7 +142,7 @@ func V1HttpClientOption() Field {
 					{
 						Label:        "Endpoint Parameters",
 						PropertyName: "endpoint_params",
-						Element:      ElementTypeKeyValueMap,
+						Element:      schema.ElementTypeKeyValueMap,
 						Description:  "Optional parameters to append to the access token request.",
 						Required:     false,
 						Secure:       false,
@@ -149,7 +151,7 @@ func V1HttpClientOption() Field {
 						Label:          "TLS",
 						PropertyName:   "tls_config",
 						Description:    "Optional TLS configuration options for OAuth2 requests.",
-						Element:        ElementTypeSubform,
+						Element:        schema.ElementTypeSubform,
 						SubformOptions: V1TLSSubformOptions(),
 					},
 					V1ProxyOption(),
